Implement skeleton code for new version of BattleManager
While flexible and easy to use, one of the main problems that I had
with the previous iteration was how heavily reliant the system was
on Unity's parent-child relationships- something I believed could be
easily broken. For example, when removing old States from the
BattleManager, it had a weird quirk where the old States would only
be removed at the end of the current frame, meaning when adding new
States, the old States would still technically be active. While not
a problem now, it could be later down the line.

Also, the BattleManager should not be a singleton and instead should
be created every time there is a new Battle.

This new version is a combination of what worked from the old version
along with some code from both the State Machine hierarchy video
I watched and the How to Make an RPG guide.