Rename Combat Queue for better understanding
When it comes to the Battle System, there are two main components:
CombatEvents and CombatStates. CombatEvents are events that occur
in battle such as Taking a Turn, Attacking, Poisoning, etc. These
will be added to the CombatEventQueue- which is ordered by time
priority. CombatStates are like the substates of these CombatEvents.
For example- Attacking- Selecting a Target, Playing the Animation of
the attack. These are the states that would have been added to the
substate list of the original BattleManager implementation.