Update logic of processing combat logic to revolve around substates rather than CombatEvents
Instead of having the confusion of processing the currentEvent
and the current substate, I decided to simplify things by
removing currentEvent. When a CombatEvent is first processed,
it is removed from the CombatEventQueue and added as the first
substate of the substate Queue. As a result, only the substate
Queue has to be worried about when processing the current CombatEvent.

Substates = current combat logic being processed.
CombatEvents = combat logic being processed in the future.