Update structure of game items to differentiate instances of same type
While ScriptableObjects are good for data containers, it creates the
issue of multiple items sharing the same data. For most
variables, this is not a problem (as that is the role of a ScriptableObject),
but it makes it harder to differentiate multiple items of the same type.

In order to account for that that, an item's data is split between
BaseItemStats and ItemStats in a similar manner to how ActorStats are structure,
with ItemStats creating a unique GUID for each item. By using this GUID, it allows
for a way to differentiate separate instances of items, even if they
are of the same type.